Fill in the Github wall once
First version of the code that actually writes to Github!

This code assumes there's not been any manipulation of your wall yet,
and generates commits for all the days that'll fill your wall.

It's only useful for filling your wall once, it's not helpful for
incremental filling over time. That comes later.